requirements.txt
attrs==22.2.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pytest==7.2.2

Appium-Python-Client~=2.8.1
selenium~=4.8.2